Add list of transient ops for GC purposes.
Care should be taken when working with fibers with regards to keeping refs to objects used in ongoing ops. Apparently, calls to RB_GC_GUARD for guarding refs on a fiber's stack are useless *unless* we keep a ref to the fiber somewhere. Otherwise, the fiber and all refs on its stack might be GC'd prematurely, leading to a segfault.

We therefore do both of the following:

1. Keep running fibers in a fibermap.
2. Keep refs to objects for transient ops in a linked list.