Fix segfault caused by waiting fibers not being marked
When a fiber waits for a completion, control is transferred to another
fiber, and the waiting fiber might be GC'd if not referenced by anything
else. This commit streamlines how fibers are referenced for
non-transient (i.e. stack-allocated) ops. The normal `um_switch`
(`um_fiber_switch` previously) performs the fiber switch automatically.
A new `um_yield` function puts the waiting fiber in the `pending_fibers`
hash, performs the switch, and then removes it when the fiber is
resumed.